Fixing issue selecting a provider without using the arrow keys
Using this project for the first time, the user experiences a issue causing the error "bad input" while selecting the identity provider.

This happens because at the first time there's no configuration file, so the value of field "Provider" in struct *IDPAccount (in *ConfigManager) is set to an empty string (the default value for this type). This field is passed as the default value for "prompter.ChooseWithDefault" inside function PromptForConfigurationDetails.

The solution proposed here, is to check the value for the field "Provider", and if it's zeroed, we can use the first item available in slice "providers".
Also, these changes are adding a similar check for *CliPrompter.ChooseWithDefault but in a more generic way, so if this happen with another field, the user won't experience this error again.